Allow passing a board-specific configuration
Building a PMUFW for the K26 SoM series requires several CFLAGS
additions. In order to simplify life to users add support to pass an
optional config parameter to trigger any board- or SoM-specific settings
via the -c flag. If the flag is not passed, the behaviour is unmodified.

Use getopts in order to have proper and future-proof parsing.

Note this commit does not add any configurations. This is done in following
commits.

Signed-off-by: Luca Ceresoli <luca@lucaceresoli.net>