toolchain: add workaround to fix reboot
The PMU Firmware built with this tool is not capable of handling a
system reboot from Linux and the reboot command ends hanging the device.
This error can be drop down to an exception in the PMU Firmware during
the ECC initialization of the TCM memories during the RPU cluster reset
process, due to the memset implementation.

Here are the detailed steps that generate the problem:

 1. During the Linux kernel boot process, to make simpler some power
    management use cases, the PMU Firmware powers down the RPU cluster if
    these are unused.

 2. When the reboot command is received by the PMU Firmware, the RPU
    cluster is reset as well as the associated TCM memories.

 3. The TCM reset process requires ECC initialization, that is implemented
    with a memset call.

 4. Both the toolchain built with Crosstool-NG and Yocto are using Newlib
    as C library but differ on the configuration used for the build. While
    the default configuration generated by crosstool-ng build script uses
    __OPTIMIZE_SIZE__, the one built with Yocto does not.

 5. The size optimized implementation of memset uses 8bit accesses to the
    TCM memory, which in fact generates a read/modify/write operation to
    the TCM, causing an ECC error in the uninitialized memory, causing an
    exception in the PMU.

In order to fix the error, the PMUFW code should be fixed to work with any
compiler. Until that happens, let's work around by compiling the C library
(newlib) with size optimization option disabled. Otherwise the optimized
version will make 8-bit accesses to the TCM generating the above mentioned
ECC error due to read/modify/write.

Signed-off-by: Ibai Erkiaga <ibai.erkiaga-elorza@amd.com>
[Luca: cleanups, improve commit message]
Signed-off-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Tested-by: Luca Ceresoli <luca.ceresoli@bootlin.com>