build.sh: merge k24 and k26 configs
AMD is about to release a new K24 SOM along with a KD240 Starter Kit.  It will
be a lower end SOM with less memory and a smaller zynqmp to support applications
that do not require the features of a K26 SOM.

Since both the K24 and K26 SOMs use the same pmufw defines, this patch
combines support for both SOMs by switching to a combined build option called
"kria".

Signed-off-by: Neal Frager <neal.frager@amd.com>