embeddedsw: bump to xilinx_v2024.1
This patch bumps the embeddedsw repo to the xilinx_v2024.1 release.

Included in the xilinx_v2024.1 release are new defines K26_SOM and K24_SOM for
building the zynqmp-pmufw for Kria SOMs.  The build.sh script has been updated
to use the new K26_SOM define.

Details of the new define can be found in the following commits:
https://github.com/Xilinx/embeddedsw/commit/b1fc62f7475c93f648468df1e40254f9f41f1995
https://github.com/Xilinx/embeddedsw/commit/b4c76e7776ceaaaf2b8f7d60b9d631bef64a451f

Signed-off-by: Neal Frager <neal.frager@amd.com>